+++ src/chatbot/services.py
@@ -12,7 +12,7 @@ from pydub import AudioSegment
 
 import src.chatbot.models as models
 from src.chatbot.database import save_group_msg_to_db
-from src.chatbot.message_templates import create_houses_carousel
+from src.chatbot.message_templates.carousel_template import create_houses_carousel
 
 line_bot_api = LineBotApi(os.getenv("CHANNEL_ACCESS_TOKEN"))
